requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
django-jazzmin==2.6.0
django-ninja==0.21.0
djangorestframework==3.14.0
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
idna==3.4
Pillow==9.5.0
pyasn1==0.4.8
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3
rsa==4.9
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.5.0
